+++ festivalApp/src/main/java/com/revature/festivalapp/services/RegInfoService.java
@@ -7,7 +7,7 @@ import com.revature.festivalapp.pojos.RegInfo;
 public interface RegInfoService {
 	
 	public void insertRegInfo(RegInfo a);
-	public RegInfo getRegInfo(int i);
+	public RegInfo getRegInfo(long i);
 	public List<RegInfo> getAllRegInfo();
 	public void updateRegInfo(RegInfo a);
 	public void deleteRegInfo(RegInfo a);
